+++ Mediabot/Plugin/InvocationV3.pm
@@ -21,6 +21,18 @@ sub _safe_scalar {
     return substr($text, 0, $limit);
 }
 
+sub _copy_config {
+    my ($config) = @_;
+    return {} unless ref($config) eq 'HASH';
+    my %copy;
+    for my $key (keys %$config) {
+        my $value = $config->{$key};
+        next if ref($value);
+        $copy{$key} = $value;
+    }
+    return \%copy;
+}
+
 sub new {
     my ($class, %args) = @_;
 
@@ -40,6 +52,11 @@ sub new {
     die "InvocationV3: PluginContext authority is required\n"
         unless ref($args{authority})
             && eval { $args{authority}->isa('Mediabot::PluginContext') };
+    my $activation = defined($args{activation}) ? $args{activation} : 'off';
+    die "InvocationV3: invalid activation mode\n"
+        unless !ref($activation) && $activation =~ /\A(?:off|observe|on)\z/;
+    die "InvocationV3: output guard must be CODE\n"
+        if exists($args{output_guard}) && ref($args{output_guard}) ne 'CODE';
 
     my $opaque = 0;
     my $self = bless \$opaque, $class;
@@ -53,6 +70,9 @@ sub new {
         reply_sink  => $args{reply_sink},
         notice_sink => $args{notice_sink},
         authority   => $args{authority},
+        activation  => "$activation",
+        config      => _copy_config($args{config}),
+        output_guard => $args{output_guard},
     };
     return $self;
 }
@@ -63,11 +83,28 @@ sub command    { _state($_[0])->{command} }
 sub source     { _state($_[0])->{source} }
 sub is_private { _state($_[0])->{is_private} ? 1 : 0 }
 sub args       { [ @{ _state($_[0])->{args} } ] }
+sub activation_mode { _state($_[0])->{activation} }
+sub config { _copy_config(_state($_[0])->{config}) }
+
+sub config_value {
+    my ($self, $key) = @_;
+    return undef unless defined($key) && !ref($key);
+    return _state($self)->{config}{$key};
+}
+
+sub output_allowed {
+    my ($self) = @_;
+    my $state = _state($self);
+    return 0 unless $state->{activation} eq 'on';
+    return 1 unless ref($state->{output_guard}) eq 'CODE';
+    return $state->{output_guard}->() ? 1 : 0;
+}
 
 sub _emit_reply {
     my ($self, $text) = @_;
     my $state = _state($self);
     $state->{authority}->require_capability('irc.reply');
+    return 0 unless $self->output_allowed;
     return $state->{reply_sink}->($text);
 }
 
@@ -75,6 +112,7 @@ sub _emit_notice {
     my ($self, $text) = @_;
     my $state = _state($self);
     $state->{authority}->require_capability('irc.notice');
+    return 0 unless $self->output_allowed;
     return $state->{notice_sink}->($text);
 }