+++ plugins/hello-v3/lib/Hello.pm
@@ -14,6 +14,9 @@ sub new {
         started => 0,
         minutes_observed => 0,
         heartbeats       => 0,
+        commands_observed => 0,
+        observed_channels => [],
+        observed_modes    => [],
     }, $class;
 }
 
@@ -32,16 +35,22 @@ sub stop {
 sub command_hello {
     my ($self, $context, $invocation) = @_;
     return unless $self->{started};
-    return $context->reply(
-        $invocation,
-        'Hello from a tiny, capability-scoped API v3 plugin.'
-    );
+    $self->{commands_observed}++;
+    push @{ $self->{observed_modes} }, $invocation->activation_mode;
+    my $greeting = $invocation->config_value('greeting');
+    my $mention = $invocation->config_value('mention_nick')
+        ? $invocation->nick . ': ' : '';
+    my $enthusiasm = $invocation->config_value('enthusiasm') // 0;
+    return $context->reply($invocation,
+        $mention . $greeting . ('!' x $enthusiasm));
 }
 
 sub event_minute {
     my ($self, $context, $event) = @_;
     return unless $self->{started};
     $self->{minutes_observed}++;
+    push @{ $self->{observed_channels} }, $event->policy_channel;
+    push @{ $self->{observed_modes} }, $event->activation_mode;
     return 1;
 }
 
@@ -49,6 +58,8 @@ sub job_heartbeat {
     my ($self, $context, $job) = @_;
     return unless $self->{started};
     $self->{heartbeats}++;
+    push @{ $self->{observed_channels} }, $job->channel;
+    push @{ $self->{observed_modes} }, $job->activation_mode;
     return 1;
 }